Log improvement.

+++ mariadb_incrementalbackup.ps1
@@ -2,13 +2,8 @@
 $Logfile = "$PSScriptRoot\mariadb_incrementalbackup.log"
 
 #### FUNCTIONS ####
-function Test-Variable {
-    param (
-        [String]$VariableName, 
-        [String]$VariableValue
-    )
-
-    if ([string]::IsNullOrEmpty($VariableValue)) {
+function Test-Variable([String]$VariableName, [String]$VariableValue){
+    if (-Not $VariableValue){
         $MessageText = "Environment variable $VariableName is not defined as expected. `
         Execute the script 'install_scripts.ps1' to set the environment variables correctly."
         Write-Error $MessageText
@@ -23,18 +18,20 @@ function WriteLog {
 }
 
 # Verify the environment variables
+WriteLog("#################################################################################")
 WriteLog("Verifying if expected Environment variables are defined...")
-Test-Variable('DB_PATH', $Env:DB_PATH)
-Test-Variable('DB_USER', $Env:DB_USER)
-Test-Variable('DB_PWD', $Env:DB_PWD)
-Test-Variable('BACKUP_PATH', $Env:BACKUP_PATH)
+Test-Variable 'DB_PATH' $Env:DB_PATH
+Test-Variable 'DB_USER' $Env:DB_USER
+Test-Variable 'DB_PWD' $Env:DB_PWD
+Test-Variable 'BACKUP_PATH' $Env:BACKUP_PATH
 
-# Checks if a backup full exists and creates one if not
+# Checks if a backup full exists
 WriteLog("Verifying if the backup full was executed previously...")
-If (-Not (Test-Path -Path "$Env:BACKUP_PATH\*")) {
+If (-Not (Test-Path -Path "$Env:BACKUP_PATH\full\*")) {
     $MessageText = "Backup full don't exists. Run backup full before incremental."
-    Write-Error $MessageText
     WriteLog($MessageText)
+    #Write-Error $MessageText
+    Throw $MessageText
 }
 
 # Check if the path "last_backup" exists and create it if it doesn't.
@@ -65,8 +62,13 @@ else {
 # Move all files to last_backup\incremental directory.
 WriteLog("Moving previous incremental backup to folder 'last_backup\incremental'...")
 try{
-    Get-ChildItem -Path "$Env:BACKUP_PATH\incremental\*" -Recurse | Move-Item -Destination "$Env:BACKUP_PATH\last_backup\incremental\"
-    WriteLog("Previous incremental backup moved successfully.")
+    If (Test-Path -Path "$Env:BACKUP_PATH\incremental\"){
+        Get-ChildItem -Path "$Env:BACKUP_PATH\incremental\*" -Recurse | Move-Item -Destination "$Env:BACKUP_PATH\last_backup\incremental\"
+        WriteLog("Previous incremental backup moved successfully.")
+    } else {
+        New-Item -Path "$Env:BACKUP_PATH\incremental\" -ItemType Directory
+        WriteLog("$Env:BACKUP_PATH\incremental\ folder created successfully")
+    }
 } catch {
     $MessageText = "Error moving previous incremental backup to folder 'last_backup\incremental'."
     Write-Error($MessageText)
